requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
aiofiles>=23.2.0

